creates string enums for IdLogicalType, CustomIdLogicalType, NonIdLogicalType. adds index signature for NonIdLogicalTypes. changes implementation of logicalTypeValueLabel to use string enums. caveat: this will only work with ts >2.6
