+++ wherehows-web/app/constants/datasets/compliance.ts
@@ -1,4 +1,12 @@
 /**
+ * String indicating that the user affirms or ignored a field suggestion
+ */
+enum SuggestionIntent {
+  accept = 'accept',
+  ignore = 'ignore'
+}
+
+/**
  * Defines the string values that are allowed for a classification
  */
 enum Classification {
@@ -7,6 +15,29 @@ enum Classification {
   HighlyConfidential = 'highlyConfidential'
 }
 
+/**
+ * Defines the string values for an id logical tyoe
+ * @enum {string}
+ */
+enum IdLogicalType {
+  Id = 'ID',
+  Urn = 'URN',
+  ReversedUrn = 'REVERSED_URN',
+  CompositeUrn = 'COMPOSITE_URN'
+}
+
+/**
+ * Defines the string values for a custom id logical tyoe
+ * @enum {string}
+ */
+enum CustomIdLogicalType {
+  Custom = 'CUSTOM_ID'
+}
+
+/**
+ * Enum of values for non id / generic logical types
+ * @enum {string}
+ */
 enum NonIdLogicalType {
   Name = 'NAME',
   Email = 'EMAIL',
@@ -28,12 +59,49 @@ enum NonIdLogicalType {
   DeviceIdAdvertisingId = 'DEVICE_ID_ADVERTISING_ID'
 }
 
+/**
+ * Describes the index signature for the nonIdFieldLogicalTypes object
+ */
 type NonIdLogicalTypesSignature = {
   [K in NonIdLogicalType]: {
     classification: Classification;
     displayAs: string;
   }
 };
+
+/**
+ * Describes the properties on a field identifier object for ui rendering
+ */
+interface IFieldIdProps {
+  value: string;
+  isId: boolean;
+  displayAs: string;
+}
+
+/**
+ * Describes the index signature for fieldIdentifierTypes
+ */
+interface IFieldIdTypes {
+  [prop: string]: IFieldIdProps;
+}
+
+/**
+ * A list of id logical types
+ * @type {Array.<String>}
+ */
+const idLogicalTypes = Object.values(IdLogicalType).sort() as Array<IdLogicalType>;
+
+/**
+ * A list of custom logical types that may be treated ids but have a different behaviour from regular ids
+ * @type {Array.<String>}
+ */
+const customIdLogicalTypes = Object.values(CustomIdLogicalType) as Array<CustomIdLogicalType>;
+
+/**
+ * List of non Id field data type classifications
+ * @type {Array}
+ */
+const genericLogicalTypes = Object.values(NonIdLogicalType).sort() as Array<NonIdLogicalType>;
 
 // Default mapping of field data types to security classification
 // https://iwww.corp.linkedin.com/wiki/cf/display/DWH/List+of+Metadata+for+Data+Sets
@@ -112,4 +180,89 @@ const nonIdFieldLogicalTypes: NonIdLogicalTypesSignature = {
   }
 };
 
-export { Classification, NonIdLogicalType, nonIdFieldLogicalTypes };
+/**
+ * A map of identifier types for fields on a dataset
+ * @type {{none: {value: string, isId: boolean, displayAs: string}, member: {value: string, isId: boolean, displayAs: string}, subjectMember: {value: string, isId: boolean, displayAs: string}, group: {value: string, isId: boolean, displayAs: string}, organization: {value: string, isId: boolean, displayAs: string}, generic: {value: string, isId: boolean, displayAs: string}}}
+ */
+const fieldIdentifierTypes: IFieldIdTypes = {
+  none: {
+    value: 'NONE',
+    isId: false,
+    displayAs: 'Not an ID'
+  },
+  member: {
+    value: 'MEMBER_ID',
+    isId: true,
+    displayAs: 'Member ID'
+  },
+  subjectMember: {
+    value: 'SUBJECT_MEMBER_ID',
+    isId: true,
+    displayAs: 'Member ID (Subject Owner)'
+  },
+  group: {
+    value: 'GROUP_ID',
+    isId: true,
+    displayAs: 'Group ID'
+  },
+  organization: {
+    value: 'COMPANY_ID',
+    isId: true,
+    displayAs: 'Organization ID'
+  },
+  generic: {
+    value: 'MIXED_ID',
+    isId: false,
+    displayAs: 'Mixed'
+  },
+  custom: {
+    value: 'CUSTOM_ID',
+    isId: false,
+    // Although rendered as though an id, it's custom and from a UI perspective does not share a key similarity to other
+    // ids, a logicalType / (field format) is not required to update this fields properties
+    displayAs: 'Custom ID'
+  },
+  enterpriseProfile: {
+    value: 'ENTERPRISE_PROFILE_ID',
+    isId: true,
+    displayAs: 'Enterprise Profile ID'
+  },
+  enterpriseAccount: {
+    value: 'ENTERPRISE_ACCOUNT_ID',
+    isId: true,
+    displayAs: 'Enterprise Account ID'
+  },
+  contract: {
+    value: 'CONTRACT_ID',
+    isId: true,
+    displayAs: 'Contract ID'
+  },
+  seat: {
+    value: 'SEAT_ID',
+    isId: true,
+    displayAs: 'Seat ID'
+  },
+  advertiser: {
+    value: 'ADVERTISER_ID',
+    isId: true,
+    displayAs: 'Advertiser ID'
+  },
+  slideshare: {
+    value: 'SLIDESHARE_USER_ID',
+    isId: true,
+    displayAs: 'SlideShare User ID'
+  }
+};
+
+export {
+  Classification,
+  NonIdLogicalType,
+  nonIdFieldLogicalTypes,
+  IFieldIdProps,
+  IdLogicalType,
+  idLogicalTypes,
+  customIdLogicalTypes,
+  genericLogicalTypes,
+  fieldIdentifierTypes,
+  SuggestionIntent
+};
