+++ wherehows-web/app/constants/index.ts
@@ -5,3 +5,4 @@ export * from 'wherehows-web/constants/application';
 export * from 'wherehows-web/constants/dataset-comments';
 export * from 'wherehows-web/constants/dataset-purge-policy';
 export * from 'wherehows-web/constants/dataset-platforms';
+export * from 'wherehows-web/constants/datasets/compliance';
