+++ wherehows-web/package.json
@@ -63,7 +63,7 @@
     "prettier": "^1.5.3",
     "redux": "^3.6.0",
     "redux-thunk": "^2.2.0",
-    "typescript": "^2.4.2"
+    "typescript": "^2.5.3"
   },
   "engines": {
     "node": ">= 0.12.0"
